+++ bob-backend/src/lib.rs
@@ -34,9 +34,9 @@ pub(crate) mod prelude {
     pub use chrono::{DateTime, Datelike, Duration as ChronoDuration, NaiveDateTime, Utc};
     pub use futures::{stream::FuturesUnordered, StreamExt, TryFutureExt};
     pub use pearl::{
-        filter::hierarchical::*, filter::traits::*, filter::Config as BloomConfig, rio, Builder,
-        Error as PearlError, ErrorKind as PearlErrorKind, Key as KeyTrait, RefKey as RefKeyTrait,
-        Storage,
+        filter::hierarchical::*, filter::traits::*, filter::Config as BloomConfig, Builder,
+        Error as PearlError, ErrorKind as PearlErrorKind, IoDriver, Key as KeyTrait,
+        RefKey as RefKeyTrait, Storage,
     };
     pub use std::{
         collections::{hash_map::Entry, HashMap},