+++ bob-backend/src/pearl/disk_controller.rs
@@ -8,6 +8,7 @@ pub(crate) mod logger;
 use crate::{core::Operation, pearl::hooks::Hooks, prelude::*};
 use logger::DisksEventsLogger;
 
+use super::holder::PearlCreationContext;
 use super::Holder;
 use super::{core::BackendResult, settings::Settings, utils::StartTimestampConfig, Group};
 
@@ -33,7 +34,6 @@ enum GroupsState {
 pub struct DiskController {
     disk: DiskPath,
     vdisks: Vec<VDiskId>,
-    dump_sem: Arc<Semaphore>,
     run_sem: Arc<Semaphore>,
     monitor_sem: Arc<Semaphore>,
     node_name: NodeName,
@@ -44,6 +44,7 @@ pub struct DiskController {
     disk_state_metric: String,
     logger: DisksEventsLogger,
     blobs_count_cached: Arc<AtomicU64>,
+    pearl_creation_context: PearlCreationContext,
 }
 
 impl DiskController {
@@ -55,13 +56,14 @@ impl DiskController {
         settings: Arc<Settings>,
         is_alien: bool,
         logger: DisksEventsLogger,
+        iodriver: IoDriver,
     ) -> Arc<Self> {
         let disk_state_metric = format!("{}.{}", DISKS_FOLDER, disk.name());
         let dump_sem = Arc::new(Semaphore::new(config.disk_access_par_degree()));
+        let pearl_creation_context = PearlCreationContext::new(dump_sem, iodriver);
         let new_dc = Self {
             disk,
             vdisks,
-            dump_sem,
             run_sem,
             monitor_sem: Arc::new(Semaphore::new(1)),
             node_name: NodeName::from(config.name()),
@@ -72,6 +74,7 @@ impl DiskController {
             disk_state_metric,
             logger,
             blobs_count_cached: Arc::new(AtomicU64::new(0)),
+            pearl_creation_context,
         };
         new_dc
             .init()
@@ -265,7 +268,7 @@ impl DiskController {
                     self.disk.name().clone(),
                     path,
                     owner_node_identifier,
-                    self.dump_sem.clone(),
+                    self.pearl_creation_context.clone(),
                 )
             })
             .collect()
@@ -276,8 +279,8 @@ impl DiskController {
         let groups = settings
             .collect_alien_groups(
                 self.disk.name(),
-                self.dump_sem.clone(),
                 &self.node_name,
+                self.pearl_creation_context.clone(),
             )
             .await?;
         trace!(
@@ -326,7 +329,7 @@ impl DiskController {
                 operation.remote_node_name().expect("Node name not found").clone(),
                 operation.vdisk_id(),
                 &self.node_name,
-                self.dump_sem.clone(),
+                self.pearl_creation_context.clone(),
             )
             .await?;
         write_lock_groups.push(group.clone());
@@ -522,7 +525,6 @@ impl DiskController {
         }
     }
 
-
     pub(crate) async fn delete(
         &self,
         op: Operation,
@@ -589,23 +591,17 @@ impl DiskController {
         }
     }
 
-
+    
     pub(crate) async fn shutdown(&self) {
         let futures = FuturesUnordered::new();
         for group in self.groups.read().await.iter() {
             let holders = group.holders();
             let holders = holders.read().await;
             for holder in holders.iter() {
-                let storage = holder.storage().read().await;
-                let storage = storage.storage().clone();
-                let id = holder.get_id();
+                let holder = holder.clone();
                 futures.push(async move {
-                    match storage.close().await {
-                        Ok(_) => debug!("holder {} closed", id),
-                        Err(e) => {
-                            error!("error closing holder{}: {} (disk: {:?})", id, e, self.disk)
-                        }
-                    }
+                    holder.close_storage().await;
+                    debug!("holder {} closed", holder.get_id());
                 });
             }
         }